requirements.txt
gensim==3.8.3
GPUtil==1.4.0
karateclub==1.0.21
matplotlib==3.3.2
networkx==2.5
node2vec==0.4.0
nodevectors==0.1.23
numpy==1.19.0
pandas==1.1.5
pyarrow==0.13.0
pyrsistent==0.16.0
scikit-learn==0.23.2
torch==1.5.1+cu101
torchvision==0.6.1+cu101
tqdm==4.50.2
dgl-cu101==0.6.1